Final commit for the exam! Added GmailSend but cent get authorized... Cant push in Saturday night because of the secret in json fail needed for newly implemented GmailSender. I need 24 hours to find out hot to undo all failed commits, rebase, put the secret in .gitignore and actually do the commit.

+++ src/main/java/com/NikolaySHA/ExclusiveService/web/controller/UserController.java
@@ -150,6 +150,7 @@ public class UserController {
             return "redirect:/users/" + id;
         }
         userService.addAdmin(id);
+        redirectAttributes.addFlashAttribute("addAdminMessage", true);
         return "redirect:/users/" + id; // Redirects to the user's page after adding admin role
     }
     
@@ -162,7 +163,11 @@ public class UserController {
             return "redirect:/error/contact-admin";
         }
         User user = optionalUser.get();
+        if (!userService.isAdmin(user.getRoles())) {
+            return "redirect:/users/" + id;
+        }
         userService.removeAdmin(id);
+        redirectAttributes.addFlashAttribute("removeAdminMessage", true);
         return "redirect:/users/" + id;
     }
